+++ clang/test/CodeGen/AArch64/neon-ldst-one.c
@@ -7,368 +7,8 @@
 
 #include <arm_neon.h>
 
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_dup_u8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0:[0-9]+]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <16 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <16 x i8> [[TMP1]], <16 x i8> [[TMP1]], <16 x i32> zeroinitializer
-// CHECK-NEXT:    ret <16 x i8> [[LANE]]
-//
-uint8x16_t test_vld1q_dup_u8(uint8_t  *a) {
-  return vld1q_dup_u8(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vld1q_dup_u16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x i16> poison, i16 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x i16> [[TMP1]], <8 x i16> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x i16> [[LANE]]
-//
-uint16x8_t test_vld1q_dup_u16(uint16_t  *a) {
-  return vld1q_dup_u16(a);
-}
-
-// CHECK-LABEL: define dso_local <4 x i32> @test_vld1q_dup_u32(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <4 x i32> poison, i32 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <4 x i32> [[TMP1]], <4 x i32> [[TMP1]], <4 x i32> zeroinitializer
-// CHECK-NEXT:    ret <4 x i32> [[LANE]]
-//
-uint32x4_t test_vld1q_dup_u32(uint32_t  *a) {
-  return vld1q_dup_u32(a);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vld1q_dup_u64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <2 x i64> poison, i64 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <2 x i64> [[TMP1]], <2 x i64> [[TMP1]], <2 x i32> zeroinitializer
-// CHECK-NEXT:    ret <2 x i64> [[LANE]]
-//
-uint64x2_t test_vld1q_dup_u64(uint64_t  *a) {
-  return vld1q_dup_u64(a);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_dup_s8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <16 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <16 x i8> [[TMP1]], <16 x i8> [[TMP1]], <16 x i32> zeroinitializer
-// CHECK-NEXT:    ret <16 x i8> [[LANE]]
-//
-int8x16_t test_vld1q_dup_s8(int8_t  *a) {
-  return vld1q_dup_s8(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vld1q_dup_s16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x i16> poison, i16 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x i16> [[TMP1]], <8 x i16> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x i16> [[LANE]]
-//
-int16x8_t test_vld1q_dup_s16(int16_t  *a) {
-  return vld1q_dup_s16(a);
-}
-
-// CHECK-LABEL: define dso_local <4 x i32> @test_vld1q_dup_s32(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <4 x i32> poison, i32 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <4 x i32> [[TMP1]], <4 x i32> [[TMP1]], <4 x i32> zeroinitializer
-// CHECK-NEXT:    ret <4 x i32> [[LANE]]
-//
-int32x4_t test_vld1q_dup_s32(int32_t  *a) {
-  return vld1q_dup_s32(a);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vld1q_dup_s64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <2 x i64> poison, i64 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <2 x i64> [[TMP1]], <2 x i64> [[TMP1]], <2 x i32> zeroinitializer
-// CHECK-NEXT:    ret <2 x i64> [[LANE]]
-//
-int64x2_t test_vld1q_dup_s64(int64_t  *a) {
-  return vld1q_dup_s64(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x half> @test_vld1q_dup_f16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load half, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x half> poison, half [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x half> [[TMP1]], <8 x half> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x half> [[LANE]]
-//
-float16x8_t test_vld1q_dup_f16(float16_t  *a) {
-  return vld1q_dup_f16(a);
-}
-
-// CHECK-LABEL: define dso_local <4 x float> @test_vld1q_dup_f32(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load float, ptr [[A]], align 4
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <4 x float> poison, float [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <4 x float> [[TMP1]], <4 x float> [[TMP1]], <4 x i32> zeroinitializer
-// CHECK-NEXT:    ret <4 x float> [[LANE]]
-//
-float32x4_t test_vld1q_dup_f32(float32_t  *a) {
-  return vld1q_dup_f32(a);
-}
-
-// CHECK-LABEL: define dso_local <2 x double> @test_vld1q_dup_f64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load double, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <2 x double> poison, double [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <2 x double> [[TMP1]], <2 x double> [[TMP1]], <2 x i32> zeroinitializer
-// CHECK-NEXT:    ret <2 x double> [[LANE]]
-//
-float64x2_t test_vld1q_dup_f64(float64_t  *a) {
-  return vld1q_dup_f64(a);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_dup_mf8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <16 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <16 x i8> [[TMP1]], <16 x i8> [[TMP1]], <16 x i32> zeroinitializer
-// CHECK-NEXT:    ret <16 x i8> [[LANE]]
-//
-mfloat8x16_t test_vld1q_dup_mf8(mfloat8_t  *a) {
-  return vld1q_dup_mf8(a);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_dup_p8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <16 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <16 x i8> [[TMP1]], <16 x i8> [[TMP1]], <16 x i32> zeroinitializer
-// CHECK-NEXT:    ret <16 x i8> [[LANE]]
-//
-poly8x16_t test_vld1q_dup_p8(poly8_t  *a) {
-  return vld1q_dup_p8(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vld1q_dup_p16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x i16> poison, i16 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x i16> [[TMP1]], <8 x i16> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x i16> [[LANE]]
-//
-poly16x8_t test_vld1q_dup_p16(poly16_t  *a) {
-  return vld1q_dup_p16(a);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vld1q_dup_p64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <2 x i64> poison, i64 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <2 x i64> [[TMP1]], <2 x i64> [[TMP1]], <2 x i32> zeroinitializer
-// CHECK-NEXT:    ret <2 x i64> [[LANE]]
-//
-poly64x2_t test_vld1q_dup_p64(poly64_t  *a) {
-  return vld1q_dup_p64(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_dup_u8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x i8> [[TMP1]], <8 x i8> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x i8> [[LANE]]
-//
-uint8x8_t test_vld1_dup_u8(uint8_t  *a) {
-  return vld1_dup_u8(a);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vld1_dup_u16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <4 x i16> poison, i16 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <4 x i16> [[TMP1]], <4 x i16> [[TMP1]], <4 x i32> zeroinitializer
-// CHECK-NEXT:    ret <4 x i16> [[LANE]]
-//
-uint16x4_t test_vld1_dup_u16(uint16_t  *a) {
-  return vld1_dup_u16(a);
-}
-
-// CHECK-LABEL: define dso_local <2 x i32> @test_vld1_dup_u32(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <2 x i32> poison, i32 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <2 x i32> [[TMP1]], <2 x i32> [[TMP1]], <2 x i32> zeroinitializer
-// CHECK-NEXT:    ret <2 x i32> [[LANE]]
-//
-uint32x2_t test_vld1_dup_u32(uint32_t  *a) {
-  return vld1_dup_u32(a);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vld1_dup_u64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <1 x i64> poison, i64 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <1 x i64> [[TMP1]], <1 x i64> [[TMP1]], <1 x i32> zeroinitializer
-// CHECK-NEXT:    ret <1 x i64> [[LANE]]
-//
-uint64x1_t test_vld1_dup_u64(uint64_t  *a) {
-  return vld1_dup_u64(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_dup_s8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x i8> [[TMP1]], <8 x i8> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x i8> [[LANE]]
-//
-int8x8_t test_vld1_dup_s8(int8_t  *a) {
-  return vld1_dup_s8(a);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vld1_dup_s16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <4 x i16> poison, i16 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <4 x i16> [[TMP1]], <4 x i16> [[TMP1]], <4 x i32> zeroinitializer
-// CHECK-NEXT:    ret <4 x i16> [[LANE]]
-//
-int16x4_t test_vld1_dup_s16(int16_t  *a) {
-  return vld1_dup_s16(a);
-}
-
-// CHECK-LABEL: define dso_local <2 x i32> @test_vld1_dup_s32(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <2 x i32> poison, i32 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <2 x i32> [[TMP1]], <2 x i32> [[TMP1]], <2 x i32> zeroinitializer
-// CHECK-NEXT:    ret <2 x i32> [[LANE]]
-//
-int32x2_t test_vld1_dup_s32(int32_t  *a) {
-  return vld1_dup_s32(a);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vld1_dup_s64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <1 x i64> poison, i64 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <1 x i64> [[TMP1]], <1 x i64> [[TMP1]], <1 x i32> zeroinitializer
-// CHECK-NEXT:    ret <1 x i64> [[LANE]]
-//
-int64x1_t test_vld1_dup_s64(int64_t  *a) {
-  return vld1_dup_s64(a);
-}
-
-// CHECK-LABEL: define dso_local <4 x half> @test_vld1_dup_f16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load half, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <4 x half> poison, half [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <4 x half> [[TMP1]], <4 x half> [[TMP1]], <4 x i32> zeroinitializer
-// CHECK-NEXT:    ret <4 x half> [[LANE]]
-//
-float16x4_t test_vld1_dup_f16(float16_t  *a) {
-  return vld1_dup_f16(a);
-}
-
-// CHECK-LABEL: define dso_local <2 x float> @test_vld1_dup_f32(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load float, ptr [[A]], align 4
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <2 x float> poison, float [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <2 x float> [[TMP1]], <2 x float> [[TMP1]], <2 x i32> zeroinitializer
-// CHECK-NEXT:    ret <2 x float> [[LANE]]
-//
-float32x2_t test_vld1_dup_f32(float32_t  *a) {
-  return vld1_dup_f32(a);
-}
-
-// CHECK-LABEL: define dso_local <1 x double> @test_vld1_dup_f64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load double, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <1 x double> poison, double [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <1 x double> [[TMP1]], <1 x double> [[TMP1]], <1 x i32> zeroinitializer
-// CHECK-NEXT:    ret <1 x double> [[LANE]]
-//
-float64x1_t test_vld1_dup_f64(float64_t  *a) {
-  return vld1_dup_f64(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_dup_mf8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x i8> [[TMP1]], <8 x i8> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x i8> [[LANE]]
-//
-mfloat8x8_t test_vld1_dup_mf8(mfloat8_t  *a) {
-  return vld1_dup_mf8(a);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_dup_p8(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <8 x i8> poison, i8 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <8 x i8> [[TMP1]], <8 x i8> [[TMP1]], <8 x i32> zeroinitializer
-// CHECK-NEXT:    ret <8 x i8> [[LANE]]
-//
-poly8x8_t test_vld1_dup_p8(poly8_t  *a) {
-  return vld1_dup_p8(a);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vld1_dup_p16(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <4 x i16> poison, i16 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <4 x i16> [[TMP1]], <4 x i16> [[TMP1]], <4 x i32> zeroinitializer
-// CHECK-NEXT:    ret <4 x i16> [[LANE]]
-//
-poly16x4_t test_vld1_dup_p16(poly16_t  *a) {
-  return vld1_dup_p16(a);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vld1_dup_p64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[TMP1:%.*]] = insertelement <1 x i64> poison, i64 [[TMP0]], i32 0
-// CHECK-NEXT:    [[LANE:%.*]] = shufflevector <1 x i64> [[TMP1]], <1 x i64> [[TMP1]], <1 x i32> zeroinitializer
-// CHECK-NEXT:    ret <1 x i64> [[LANE]]
-//
-poly64x1_t test_vld1_dup_p64(poly64_t  *a) {
-  return vld1_dup_p64(a);
-}
-
 // CHECK-LABEL: define dso_local %struct.uint64x2x2_t @test_vld2q_dup_u64(
-// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0]] {
+// CHECK-SAME: ptr noundef [[A:%.*]]) #[[ATTR0:[0-9]+]] {
 // CHECK-NEXT:  [[ENTRY:.*:]]
 // CHECK-NEXT:    [[VLD2:%.*]] = call { <2 x i64>, <2 x i64> } @llvm.aarch64.neon.ld2r.v2i64.p0(ptr [[A]])
 // CHECK-NEXT:    [[VLD2_FCA_0_EXTRACT:%.*]] = extractvalue { <2 x i64>, <2 x i64> } [[VLD2]], 0
@@ -755,387 +395,6 @@ mfloat8x8x4_t test_vld4_dup_mf8(mfloat8_t  *a) {
 //
 poly64x1x4_t test_vld4_dup_p64(poly64_t  *a) {
   return vld4_dup_p64(a);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_lane_u8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <16 x i8> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <16 x i8> [[B]], i8 [[TMP0]], i32 15
-// CHECK-NEXT:    ret <16 x i8> [[VLD1_LANE]]
-//
-uint8x16_t test_vld1q_lane_u8(uint8_t  *a, uint8x16_t b) {
-  return vld1q_lane_u8(a, b, 15);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vld1q_lane_u16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x i16> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <8 x i16> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <8 x i16>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x i16> [[TMP1]], i16 [[TMP2]], i32 7
-// CHECK-NEXT:    ret <8 x i16> [[VLD1_LANE]]
-//
-uint16x8_t test_vld1q_lane_u16(uint16_t  *a, uint16x8_t b) {
-  return vld1q_lane_u16(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <4 x i32> @test_vld1q_lane_u32(
-// CHECK-SAME: ptr noundef [[A:%.*]], <4 x i32> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <4 x i32> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <4 x i32>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <4 x i32> [[TMP1]], i32 [[TMP2]], i32 3
-// CHECK-NEXT:    ret <4 x i32> [[VLD1_LANE]]
-//
-uint32x4_t test_vld1q_lane_u32(uint32_t  *a, uint32x4_t b) {
-  return vld1q_lane_u32(a, b, 3);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vld1q_lane_u64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <2 x i64> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <2 x i64> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <2 x i64>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <2 x i64> [[TMP1]], i64 [[TMP2]], i32 1
-// CHECK-NEXT:    ret <2 x i64> [[VLD1_LANE]]
-//
-uint64x2_t test_vld1q_lane_u64(uint64_t  *a, uint64x2_t b) {
-  return vld1q_lane_u64(a, b, 1);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_lane_s8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <16 x i8> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <16 x i8> [[B]], i8 [[TMP0]], i32 15
-// CHECK-NEXT:    ret <16 x i8> [[VLD1_LANE]]
-//
-int8x16_t test_vld1q_lane_s8(int8_t  *a, int8x16_t b) {
-  return vld1q_lane_s8(a, b, 15);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vld1q_lane_s16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x i16> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <8 x i16> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <8 x i16>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x i16> [[TMP1]], i16 [[TMP2]], i32 7
-// CHECK-NEXT:    ret <8 x i16> [[VLD1_LANE]]
-//
-int16x8_t test_vld1q_lane_s16(int16_t  *a, int16x8_t b) {
-  return vld1q_lane_s16(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <4 x i32> @test_vld1q_lane_s32(
-// CHECK-SAME: ptr noundef [[A:%.*]], <4 x i32> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <4 x i32> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <4 x i32>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <4 x i32> [[TMP1]], i32 [[TMP2]], i32 3
-// CHECK-NEXT:    ret <4 x i32> [[VLD1_LANE]]
-//
-int32x4_t test_vld1q_lane_s32(int32_t  *a, int32x4_t b) {
-  return vld1q_lane_s32(a, b, 3);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vld1q_lane_s64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <2 x i64> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <2 x i64> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <2 x i64>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <2 x i64> [[TMP1]], i64 [[TMP2]], i32 1
-// CHECK-NEXT:    ret <2 x i64> [[VLD1_LANE]]
-//
-int64x2_t test_vld1q_lane_s64(int64_t  *a, int64x2_t b) {
-  return vld1q_lane_s64(a, b, 1);
-}
-
-// CHECK-LABEL: define dso_local <8 x half> @test_vld1q_lane_f16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x half> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <8 x half> [[B]] to <8 x i16>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i16> [[TMP0]] to <16 x i8>
-// CHECK-NEXT:    [[TMP2:%.*]] = bitcast <16 x i8> [[TMP1]] to <8 x half>
-// CHECK-NEXT:    [[TMP3:%.*]] = load half, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x half> [[TMP2]], half [[TMP3]], i32 7
-// CHECK-NEXT:    ret <8 x half> [[VLD1_LANE]]
-//
-float16x8_t test_vld1q_lane_f16(float16_t  *a, float16x8_t b) {
-  return vld1q_lane_f16(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <4 x float> @test_vld1q_lane_f32(
-// CHECK-SAME: ptr noundef [[A:%.*]], <4 x float> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <4 x float> [[B]] to <4 x i32>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <4 x i32> [[TMP0]] to <16 x i8>
-// CHECK-NEXT:    [[TMP2:%.*]] = bitcast <16 x i8> [[TMP1]] to <4 x float>
-// CHECK-NEXT:    [[TMP3:%.*]] = load float, ptr [[A]], align 4
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <4 x float> [[TMP2]], float [[TMP3]], i32 3
-// CHECK-NEXT:    ret <4 x float> [[VLD1_LANE]]
-//
-float32x4_t test_vld1q_lane_f32(float32_t  *a, float32x4_t b) {
-  return vld1q_lane_f32(a, b, 3);
-}
-
-// CHECK-LABEL: define dso_local <2 x double> @test_vld1q_lane_f64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <2 x double> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <2 x double> [[B]] to <2 x i64>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <2 x i64> [[TMP0]] to <16 x i8>
-// CHECK-NEXT:    [[TMP2:%.*]] = bitcast <16 x i8> [[TMP1]] to <2 x double>
-// CHECK-NEXT:    [[TMP3:%.*]] = load double, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <2 x double> [[TMP2]], double [[TMP3]], i32 1
-// CHECK-NEXT:    ret <2 x double> [[VLD1_LANE]]
-//
-float64x2_t test_vld1q_lane_f64(float64_t  *a, float64x2_t b) {
-  return vld1q_lane_f64(a, b, 1);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_lane_mf8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <16 x i8> [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <16 x i8> [[B]], i8 [[TMP0]], i32 15
-// CHECK-NEXT:    ret <16 x i8> [[VLD1_LANE]]
-//
-mfloat8x16_t test_vld1q_lane_mf8(mfloat8_t  *a, mfloat8x16_t b) {
-  return vld1q_lane_mf8(a, b, 15);
-}
-
-// CHECK-LABEL: define dso_local <16 x i8> @test_vld1q_lane_p8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <16 x i8> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <16 x i8> [[B]], i8 [[TMP0]], i32 15
-// CHECK-NEXT:    ret <16 x i8> [[VLD1_LANE]]
-//
-poly8x16_t test_vld1q_lane_p8(poly8_t  *a, poly8x16_t b) {
-  return vld1q_lane_p8(a, b, 15);
-}
-
-// CHECK-LABEL: define dso_local <8 x i16> @test_vld1q_lane_p16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x i16> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <8 x i16> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <8 x i16>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x i16> [[TMP1]], i16 [[TMP2]], i32 7
-// CHECK-NEXT:    ret <8 x i16> [[VLD1_LANE]]
-//
-poly16x8_t test_vld1q_lane_p16(poly16_t  *a, poly16x8_t b) {
-  return vld1q_lane_p16(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <2 x i64> @test_vld1q_lane_p64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <2 x i64> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <2 x i64> [[B]] to <16 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <16 x i8> [[TMP0]] to <2 x i64>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <2 x i64> [[TMP1]], i64 [[TMP2]], i32 1
-// CHECK-NEXT:    ret <2 x i64> [[VLD1_LANE]]
-//
-poly64x2_t test_vld1q_lane_p64(poly64_t  *a, poly64x2_t b) {
-  return vld1q_lane_p64(a, b, 1);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_lane_u8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x i8> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x i8> [[B]], i8 [[TMP0]], i32 7
-// CHECK-NEXT:    ret <8 x i8> [[VLD1_LANE]]
-//
-uint8x8_t test_vld1_lane_u8(uint8_t  *a, uint8x8_t b) {
-  return vld1_lane_u8(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vld1_lane_u16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <4 x i16> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <4 x i16> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <4 x i16>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <4 x i16> [[TMP1]], i16 [[TMP2]], i32 3
-// CHECK-NEXT:    ret <4 x i16> [[VLD1_LANE]]
-//
-uint16x4_t test_vld1_lane_u16(uint16_t  *a, uint16x4_t b) {
-  return vld1_lane_u16(a, b, 3);
-}
-
-// CHECK-LABEL: define dso_local <2 x i32> @test_vld1_lane_u32(
-// CHECK-SAME: ptr noundef [[A:%.*]], <2 x i32> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <2 x i32> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <2 x i32>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <2 x i32> [[TMP1]], i32 [[TMP2]], i32 1
-// CHECK-NEXT:    ret <2 x i32> [[VLD1_LANE]]
-//
-uint32x2_t test_vld1_lane_u32(uint32_t  *a, uint32x2_t b) {
-  return vld1_lane_u32(a, b, 1);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vld1_lane_u64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <1 x i64> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <1 x i64> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <1 x i64>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <1 x i64> [[TMP1]], i64 [[TMP2]], i32 0
-// CHECK-NEXT:    ret <1 x i64> [[VLD1_LANE]]
-//
-uint64x1_t test_vld1_lane_u64(uint64_t  *a, uint64x1_t b) {
-  return vld1_lane_u64(a, b, 0);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_lane_s8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x i8> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x i8> [[B]], i8 [[TMP0]], i32 7
-// CHECK-NEXT:    ret <8 x i8> [[VLD1_LANE]]
-//
-int8x8_t test_vld1_lane_s8(int8_t  *a, int8x8_t b) {
-  return vld1_lane_s8(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vld1_lane_s16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <4 x i16> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <4 x i16> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <4 x i16>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <4 x i16> [[TMP1]], i16 [[TMP2]], i32 3
-// CHECK-NEXT:    ret <4 x i16> [[VLD1_LANE]]
-//
-int16x4_t test_vld1_lane_s16(int16_t  *a, int16x4_t b) {
-  return vld1_lane_s16(a, b, 3);
-}
-
-// CHECK-LABEL: define dso_local <2 x i32> @test_vld1_lane_s32(
-// CHECK-SAME: ptr noundef [[A:%.*]], <2 x i32> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <2 x i32> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <2 x i32>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i32, ptr [[A]], align 4
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <2 x i32> [[TMP1]], i32 [[TMP2]], i32 1
-// CHECK-NEXT:    ret <2 x i32> [[VLD1_LANE]]
-//
-int32x2_t test_vld1_lane_s32(int32_t  *a, int32x2_t b) {
-  return vld1_lane_s32(a, b, 1);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vld1_lane_s64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <1 x i64> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <1 x i64> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <1 x i64>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <1 x i64> [[TMP1]], i64 [[TMP2]], i32 0
-// CHECK-NEXT:    ret <1 x i64> [[VLD1_LANE]]
-//
-int64x1_t test_vld1_lane_s64(int64_t  *a, int64x1_t b) {
-  return vld1_lane_s64(a, b, 0);
-}
-
-// CHECK-LABEL: define dso_local <4 x half> @test_vld1_lane_f16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <4 x half> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <4 x half> [[B]] to <4 x i16>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <4 x i16> [[TMP0]] to <8 x i8>
-// CHECK-NEXT:    [[TMP2:%.*]] = bitcast <8 x i8> [[TMP1]] to <4 x half>
-// CHECK-NEXT:    [[TMP3:%.*]] = load half, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <4 x half> [[TMP2]], half [[TMP3]], i32 3
-// CHECK-NEXT:    ret <4 x half> [[VLD1_LANE]]
-//
-float16x4_t test_vld1_lane_f16(float16_t  *a, float16x4_t b) {
-  return vld1_lane_f16(a, b, 3);
-}
-
-// CHECK-LABEL: define dso_local <2 x float> @test_vld1_lane_f32(
-// CHECK-SAME: ptr noundef [[A:%.*]], <2 x float> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <2 x float> [[B]] to <2 x i32>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <2 x i32> [[TMP0]] to <8 x i8>
-// CHECK-NEXT:    [[TMP2:%.*]] = bitcast <8 x i8> [[TMP1]] to <2 x float>
-// CHECK-NEXT:    [[TMP3:%.*]] = load float, ptr [[A]], align 4
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <2 x float> [[TMP2]], float [[TMP3]], i32 1
-// CHECK-NEXT:    ret <2 x float> [[VLD1_LANE]]
-//
-float32x2_t test_vld1_lane_f32(float32_t  *a, float32x2_t b) {
-  return vld1_lane_f32(a, b, 1);
-}
-
-// CHECK-LABEL: define dso_local <1 x double> @test_vld1_lane_f64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <1 x double> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <1 x double> [[B]] to i64
-// CHECK-NEXT:    [[__S1_SROA_0_0_VEC_INSERT:%.*]] = insertelement <1 x i64> undef, i64 [[TMP0]], i64 0
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <1 x i64> [[__S1_SROA_0_0_VEC_INSERT]] to <8 x i8>
-// CHECK-NEXT:    [[TMP2:%.*]] = bitcast <8 x i8> [[TMP1]] to <1 x double>
-// CHECK-NEXT:    [[TMP3:%.*]] = load double, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <1 x double> [[TMP2]], double [[TMP3]], i32 0
-// CHECK-NEXT:    ret <1 x double> [[VLD1_LANE]]
-//
-float64x1_t test_vld1_lane_f64(float64_t  *a, float64x1_t b) {
-  return vld1_lane_f64(a, b, 0);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_lane_mf8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x i8> [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x i8> [[B]], i8 [[TMP0]], i32 7
-// CHECK-NEXT:    ret <8 x i8> [[VLD1_LANE]]
-//
-mfloat8x8_t test_vld1_lane_mf8(mfloat8_t  *a, mfloat8x8_t b) {
-  return vld1_lane_mf8(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <8 x i8> @test_vld1_lane_p8(
-// CHECK-SAME: ptr noundef [[A:%.*]], <8 x i8> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = load i8, ptr [[A]], align 1
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <8 x i8> [[B]], i8 [[TMP0]], i32 7
-// CHECK-NEXT:    ret <8 x i8> [[VLD1_LANE]]
-//
-poly8x8_t test_vld1_lane_p8(poly8_t  *a, poly8x8_t b) {
-  return vld1_lane_p8(a, b, 7);
-}
-
-// CHECK-LABEL: define dso_local <4 x i16> @test_vld1_lane_p16(
-// CHECK-SAME: ptr noundef [[A:%.*]], <4 x i16> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <4 x i16> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <4 x i16>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i16, ptr [[A]], align 2
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <4 x i16> [[TMP1]], i16 [[TMP2]], i32 3
-// CHECK-NEXT:    ret <4 x i16> [[VLD1_LANE]]
-//
-poly16x4_t test_vld1_lane_p16(poly16_t  *a, poly16x4_t b) {
-  return vld1_lane_p16(a, b, 3);
-}
-
-// CHECK-LABEL: define dso_local <1 x i64> @test_vld1_lane_p64(
-// CHECK-SAME: ptr noundef [[A:%.*]], <1 x i64> noundef [[B:%.*]]) #[[ATTR0]] {
-// CHECK-NEXT:  [[ENTRY:.*:]]
-// CHECK-NEXT:    [[TMP0:%.*]] = bitcast <1 x i64> [[B]] to <8 x i8>
-// CHECK-NEXT:    [[TMP1:%.*]] = bitcast <8 x i8> [[TMP0]] to <1 x i64>
-// CHECK-NEXT:    [[TMP2:%.*]] = load i64, ptr [[A]], align 8
-// CHECK-NEXT:    [[VLD1_LANE:%.*]] = insertelement <1 x i64> [[TMP1]], i64 [[TMP2]], i32 0
-// CHECK-NEXT:    ret <1 x i64> [[VLD1_LANE]]
-//
-poly64x1_t test_vld1_lane_p64(poly64_t  *a, poly64x1_t b) {
-  return vld1_lane_p64(a, b, 0);
 }
 
 // CHECK-LABEL: define dso_local %struct.int8x16x2_t @test_vld2q_lane_s8(
@@ -2010,7 +1269,6 @@ float64x2x3_t test_vld3q_lane_f64(float64_t  *a, float64x2x3_t b) {
 mfloat8x16x3_t test_vld3q_lane_mf8(mfloat8_t  *a, mfloat8x16x3_t b) {
   return vld3q_lane_mf8(a, b, 15);
 }
-
 
 // CHECK-LABEL: define dso_local %struct.poly8x16x3_t @test_vld3q_lane_p8(
 // CHECK-SAME: ptr noundef [[A:%.*]], [3 x <16 x i8>] alignstack(16) [[B_COERCE:%.*]]) #[[ATTR0]] {
